libs/board/src/lib/Board.spec.ts
import { Board } from './Board';
import { Element } from './Element';

describe('Board', () => {
  let board: Board;
  const title = 'Test Board';
  const description = 'This is a test board';

  beforeEach(() => {
    board = new Board(title, description);
  });

  describe('constructor', () => {
    it('initializes with the given title and description', () => {
      // Assuming TBoard stores title and description in a way accessible for verification
      expect(board.title).toBe(title);
      expect(board.description).toBe(description);
    });

    it('initializes with an empty children array', () => {
      expect(board.children).toEqual([]);
    });
  });

  describe('children property', () => {
    it('sets and gets children correctly', () => {
      const children = [new Element(), new Element()];
      board.children = children;
      expect(board.children).toBe(children);
    });
  });

  describe('start method', () => {
    it('should be callable without errors', () => {
      expect(() => {
        board.start();
      }).not.toThrow();
    });
  });

  describe('update method', () => {
    it('should be callable with a delta time without errors', () => {
      expect(() => {
        board.update(1);
      }).not.toThrow();
    });
  });

  describe('destroy method', () => {
    it('should be callable without errors', () => {
      expect(() => {
        board.destroy();
      }).not.toThrow();
    });
  });

  // Additional tests could be added here if start, update, and destroy methods are implemented with more specific logic.
});
